+++ database/migrations/2023_11_02_205627_create_movement_pretty_cashes_table.php
@@ -16,7 +16,7 @@ class CreateMovementPrettyCashesTable extends Migration
         Schema::create('movement_pretty_cashes', function (Blueprint $table) {
             $table->id();
 
-            $table->integer('pretty_cash_id')->unsigned();            
+            $table->unsignedBigInteger('pretty_cash_id')->unsigned();            
             $table->foreign('pretty_cash_id')->references('id')->on('pretty_cashes');
 
             $table->date('date');
